Clean up

+++ demo-2/src/main/java/com/me/dataentry/repo/StateRepo.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import com.me.dataentry.entity.State;
 
-public interface StateRepo extends PagingAndSortingRepository<State, BigDecimal> {
+public interface StateRepo extends PagingAndSortingRepository<State, Long> {
 
 }
